requirements.txt
Flask
openai
python-dotenv
requests
line-bot-sdk
